feat: Generation of Activity data
This is the first commit of the ActivityGenerator.
The ActivityGenerator is now able to generate a day worth of activity data in 5 minute intervals.
Every 5 minute interval will have some information on steps taken (if any), standHrs, and calories burned.
Data generated is to be used as a simulation of user's data coming in from sensors in wearable tech for example.