test: Added initial test for activity model
models_init_test.go was added, in which the TestMain for the models
package will live. The TestMain will be responsible for initializing
our connections or any other test setups.
The first test added lives in the activity_test.go file. This first
test is just testing our reserveTimes function to ensure that we
are allocating the right amount of hours for a given activity.
Future activity tests will live under the same file: activity_test.go