chore: Updated circleci config to test, build, and publish image
The circleci/config yaml file has been updated to be able to test,
build, and publish the ActivityGenerator image to the registry.
Each workflow has specific times when they will run.
For example, the test workflow will run on pushes to the repo, while
build and publish will only happen when there is a new release (tagged).